ref/boostrap-iso: add inputs to iso store contents
There was never an intention to setup the registry with aliases, which
would not avoid a network call ayhow if an input is not specified as an
registry-resolvable input.

At the same time, this clarification also eliniates an error that is
addressed upstream via https://github.com/NixOS/nixpkgs/pull/131814

---

work-around: for spurious reasons w.r.t. functor attrs in module system

Without any obvious reason, the module system appears to substitute
attrs that contain a `__functor` with the value of that functor.