fix: bootstrap semantics
prior to this commit the bootstrapping iso deactivated _all_ profiles
that were somehow contained in a suite.

that is an error, since it deactivates all profiles, even the most basic
and desired ones.

the impetus had been to prevent unwantes systemd services from launching

however, there seems no reliable approach to achieve that.

the now proposed alternative model is to add a bootsrap host akin to
the followint config:

```nix
# os/hosts/bootstrap.nix
{ profiles, ... }:
{
 imports = [
    profiles.networking
    profiles.users.root
    profiles.users.nixos
  ];
}
```